+++ lib/core/logger/logger_providers.dart
@@ -4,8 +4,6 @@ import 'package:talker/talker.dart'
     show Talker, TalkerKey, TalkerSettings, TimeFormat;
 
 final logger = Talker(
-  // add riverpod observer when support version 3.0.0+
-  // observer: null,
   settings: TalkerSettings(
     timeFormat: TimeFormat.yearMonthDayAndTime,
     titles: {TalkerKey.error: 'Pudding error'},