+++ lib/features/puddings/presentation/puddings_screen.dart
@@ -1,9 +1,11 @@
 import 'package:cloud_firestore/cloud_firestore.dart' show Timestamp;
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:pudding/core/components/full_page_loading.dart';
 import 'package:pudding/core/components/page_view_wrappers.dart';
 import 'package:pudding/core/logger/logger_providers.dart' show logger;
 import 'package:pudding/core/models/appbar_cfg_model.dart';
+import 'package:pudding/core/utils/routing.dart';
 import 'package:pudding/features/puddings/data/pudding_model.dart';
 import 'package:pudding/features/puddings/providers/puddings_providers.dart';
 
@@ -40,8 +42,9 @@ class PuddingsScreen extends ConsumerWidget {
             );
           },
         ),
-        loading: () => const Center(child: CircularProgressIndicator()),
-        error: (err, stack) => Center(child: Text('Error: $err')),
+        loading: () => FullPageLoading.df,
+        error: (err, stack) =>
+            errorPageWrapper(e: "loading puddings error: $err"),
       ),
       floatingActionButton: FloatingActionButton(
         onPressed: () => _addPuddingDialog(context, ref),
@@ -72,7 +75,7 @@ class PuddingsScreen extends ConsumerWidget {
         ),
         actions: [
           TextButton(
-            onPressed: () => Navigator.of(context).pop(),
+            onPressed: () => Routing.popPage(context),
             child: const Text('Cancel'),
           ),
           ElevatedButton(
@@ -88,7 +91,7 @@ class PuddingsScreen extends ConsumerWidget {
                 ref
                     .read(puddingsRepositoryProvider)
                     .addPudding(collectionId, newPudding);
-                Navigator.of(context).pop();
+                Routing.popPage(context);
                 logger.info("new pudding added");
               }
             },