+++ test/widget_test.dart
@@ -11,9 +11,9 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:pudding/main.dart';
 
 void main() {
-  testWidgets('Counter increments smoke test', (WidgetTester tester) async {
+  testWidgets('Pudding app test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const Pudding());
+    await tester.pumpWidget(const PuddingApp());
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);